modularizing methods
